backend/pyproject.toml
[project]
name = "project-experiment"
version = "1.0.0"
description = ""
authors = [
    {name = "u-hossy",email = "95125392+u-hossy@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]

[tool.poetry]
packages = [{ include = "core" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
